jax.numpy: move linspace & friends into array_creation
