+++ jax/numpy/__init__.py
@@ -93,7 +93,6 @@ from jax._src.numpy.lax_numpy import (
     fromstring as fromstring,
     from_dlpack as from_dlpack,
     gcd as gcd,
-    geomspace as geomspace,
     get_printoptions as get_printoptions,
     gradient as gradient,
     histogram as histogram,
@@ -118,9 +117,7 @@ from jax._src.numpy.lax_numpy import (
     ix_ as ix_,
     kron as kron,
     lcm as lcm,
-    linspace as linspace,
     load as load,
-    logspace as logspace,
     mask_indices as mask_indices,
     matrix_transpose as matrix_transpose,
     meshgrid as meshgrid,
@@ -180,6 +177,9 @@ from jax._src.numpy.array_creation import (
     empty_like as empty_like,
     full as full,
     full_like as full_like,
+    geomspace as geomspace,
+    linspace as linspace,
+    logspace as logspace,
     ones as ones,
     ones_like as ones_like,
     zeros as zeros,
